Updated juno to include fix for CVE-2015-3241 - 26 Aug 2015
Updates all repo SHAs to include:
 - [OSSA 2015-015] Nova instance migration process does not stop when
   instance is deleted (CVE-2015-3241)
 - Cannot rebuild a VM created from a Cinder volume backed by NetApp
   https://review.openstack.org/203253

This patch removes all SHA specifications for the OpenStack clients
so that any client requirement versions are determined solely from
the requirements of the OpenStack service and Global requirements
files.

This resolves issues where the client versions we carry are
incompatible with the services, resulting in unexpected failures.

Tempest requirements are specifically ignored as the tempest role
downloads its required clients directly from pip.

Finally, due to the change in the openstack-client version used the
heat domain/role/user tasks fail as they require the CLI to specify
the user and project domain name/id. This patch includes the CLI
changes to ensure that the deployment completes properly.

Change-Id: I88e705d063575f55c32350ead25bc7cc0bcdec08
Closes-Bug: #1489947
Closes-Bug: #1488315
Closes-Bug: #1400881