Set the Ubuntu mirror used based on the environment
Apt updates/installs are the most common cause of build timeouts
within HP Cloud gate checks.

This patch changes both the AIO and the gate check from using the
Rackspace Ubuntu mirror for everything to:

1) The AIO bootstrap will use whichever mirror is already defined
   on the host operating system.

2) The gate check will use a known mirror which is defined per
   cloud provider. If the cloud provider is not known, or the
   script is being used outside of OpenStack-CI, then the gate
   checkwill fall back to using the AIO's mechanism.

The idea is to make use of an apt mirror that is as close to the
instance as possible in order to speed up apt updates and installs.

Closes-Bug: #1491749
Change-Id: Ia37100bc198b64b74f27ea98ec5956e7cf084883
(cherry picked from commit b940e7b7de75bb0842ee36330e69103a98bd3a14)