Used named veth pairs that match container
The default veth names from LXC make it difficult to tell which veth
corresponds to each container.  This patch sets a unique veth name
that matches the container as well as the network device inside the
container.  It should make troubleshooting a little easier.

Oddly enough, this patch seems to fix or greatly reduce the occurence
of the issues seen in ticket 1491440.

Partial-Bug: 1491440
Implements: blueprint named-veths
Change-Id: I28553fd1b4f36991e11d55d56c3f0f46af9e52be