Set client versions deployed to use global requirements
This patch removes all SHA specifications for the OpenStack clients
so that any client requirement versions are determined solely from
the requirements of the OpenStack service and Global requirements
files.

This resolves issues where the client versions we carry are
incompatible with the services, resulting in unexpected failures.

The Tempest SHA configuration is moved into openstack_other so that
it can be addressed seperately when updating the SHA's. This will
be useful when master is branched to become liberty and all the
OpenStack services start to track an upstream stable/liberty branch.

The Global Requirements SHA configuration is shifted to
openstack_services for a similar reason. It will track an upstream
stable/liberty branch with all the openstack services.

Change-Id: I8f90389d22d2581814c4c62cfd37d0a4405bfd30