Set client versions deployed to use global requirements
This patch removes all SHA specifications for the OpenStack clients
so that any client requirement versions are determined solely from
the requirements of the OpenStack service and Global requirements
files.

This resolves issues where the client versions we carry are
incompatible with the services, resulting in unexpected failures.

The Tempest SHA configuration is moved into openstack_other so that
it can be addressed seperately when updating the SHA's.

The Global Requirements SHA configuration is shifted to
openstack_services because it tracks the upstream stable/kilo branch.

Change-Id: I832fa00999547cb385bebf44cfc158a61daf0e4f