Create config_template plugin
The change introduces the `config_template` plugin within our ansible
source tree with the intention to replace the `copy_update` module with
a proper ansible action plugin.

The new module implements a method to overwrite and add items into
configuration files without the need to specify all items in templates
as default variables.

The plugin currently supports ini, json, and yaml files and can be
a drop in replacement for the current template and copy_update
usage as it pertains to the various configuration files that we need
for OpenStack.

Partially implements: blueprint tunable-openstack-configuration

Change-Id: Ice1d3e6b99976c0df780f2a1e9c3e6244c3e0114