Allow protocol to be set per endpoint-type
If services are running behind an SSL terminating LB you will want to
differentiate between protocol on internalURL and publicURL endpoints.
This patch allows you to set the values of protocol per endpoint type,
but doesn't change the default behaviour which is to have it set in one
var.

The protocol for each endpoint-type can also be set globally for all
services, e.g. internaluri can be http, and publicuri can be https
through the variables:
  - openstack_service_publicuri_proto
  - openstack_service_adminuri_proto
  - openstack_service_internaluri_proto

The following variables have been renamed for forward compatibility when
upgrading to Kilo:
  - registry_host -> glance_registry_host
  - auth_protocol -> keystone_service_proto
  - auth_public_port -> keystone_service_port
  - auth_port -> keystone_admin_port

Configuration of service ports has also been made more consistent.

Based on commits:
1a2907c6a137bc4bc110f971f34904a20a7a5ad4
2b6b896105c7d4cd3b615432da3c084fcc2bf29b

Change-Id: I9256ca9d6ffedd61e12ed7b80916f491e6ecac67
Closes-Bug: #1399383