Set affinity to 1 for OpenStack-CI gate checks
This change sets the affinity for the RabbitMQ, Galera, Repo, Keystone and
Horizon containers to 1 when running on OpenStack-CI. The intention here is
to reduce the number of failures we see within the gate due to resource
starvation.

This patch is specifically only going into the master (Liberty) branch as
this is the only place where resource starvation has become critical. This
patch is a holdover until the current gate check design is revised after the
Liberty release. The redesign will split the current monolithic gate check
into multiple use-cases which each use less resources. This is described in
https://review.openstack.org/221009

Co-Authored-By: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>
Closes-Bug: #1498009
Change-Id: If159c75428f07e7f92c6c1807a054feca6292ad5