Adjust default Keystone httpd processes and threads
This patch sets the Apache threads and processes for Keystone to match
the recommendations based on performance testing done by the Keystone
core team.

This resolves an issue where the cryptography library will not allow
multiple threads to register the osrandom engine.

To compensate for the reduction in threads, the number of processes has
been increased. Both the values remain tunable to suit a deployers needs.

Change-Id: I1e94a91642b904b865ddf56983239838ed9a15e6
Closes-Bug: #1497283