Temporarily pins to unblock gating
This also addresses an issue in the openstackclient where the tagged
version has a neutron client version dependency which is unresolvable
without the client entry being presnet to allow the repo infra to
index the requirements.

Version 2.0.0 of the django openstack auth package was recently
released and causes session backed with mysql for horizon to fail a
DB sync.

Change-Id: Icd18f77e2b1525d4769c3df604030e7aa73917f3