Add net.netfilter.nf_conntrack_max to Swift Storage
With the default sysctl value, the nf_contract table fills and starts
dropping packets on Swift storage nodes after a certain period of time.

This is not a problem in Kilo as the value is set to 256k in all hosts
by default. Adding this specifically to the storage setup to avoid
adding the more complex solution used for nova/neutron that uses var
files in playbooks.

Change-Id: Ic9162eeb50523b32f477075b565f55bbf868d1d6
Closes-Bug: #1451217
Closes-Bug: #1479127