Put horizon in its own process
This change makes mod_wsgi create standalone processes for horizon. This
avoids obscure start up errors when mod_wsgi needs to spawn a new
process or thread. These restarts occur, for example, when mod_wsgi
receives a request from a client that is accessing the service through a
non-standard SSL port that is mapped to the port mod_wsgi is listening
on.

Change-Id: I5907b5c06cef0272ee145c277f79f23ba37741a4
Closes-Bug: #1499073