Add ebtables to neutron agent configuration
Neutron now uses ebtables as an extra security layer for ARP spoof filtering.
This patch adds the ebtables package and rootwrap to the neutron role to
ensure that the agent is able to use this subsystem. Without it the networking
from the instances to the L3 router will fail.

The neutron agent prevent_arp_spoofing conf option is configurable, but is set
with the default value of False to match upstream and also to prevent a change
in behaviour from previous Juno deployments.

Co-Authored-By: Evan Callicoat <diopter@gmail.com>
Closes-Bug: #1482756
Change-Id: Ibc960564a3acfbb10cfbc3cfe0ad60d3366d2443