Allow pip extra arguments to be passed to playbooks
This patch adds the variable 'pip_install_options' which is passed to the pip
install module as extra arguments in order to allow the use of options like
'--force-reinstall' when executing playbooks.

eg: openstack-ansible -e pip_install_options="--force-reinstall" \
        setup-openstack.yml

This is required due to constant upstream changes in dependencies which
result in python wheel version upgrades and downgrades between tagged
versions of openstack-ansible.

The intention is that this can be used whenever a deployer switches between
tags for both upgrades and downgrades.

DocImpact
Closes-Bug: #1489251
Closes-Bug: #1499451
Related-Bug: #1501114
Change-Id: I996185e009a4c4af4f23798619bdbd0d490360c9