Limit the number of Ansible forks used to 10
The default MaxSessions setting for SSHD is 10. Each Ansible fork makes use of
a Session, so this patch still uses the CPU number to set the number of forks
used but limits it to 10 forks when the number of CPU's is larger.

Developer Docs and Install Guide Docs entries have been included.

Closes-Bug: #1479812
Change-Id: I9abd33e184c706796ede9963393876a8aae9837c
(cherry picked from commit fe3b328023b946413a675a0ffdbafcad353aca44)