Implement a fall back URL for get-pip.py
This patch implements a fall-back URL for get-pip.py to cater for situations
where https://bootstrap.pypa.io/get-pip.py is not available (like today).

The deployer may still override the URL for the bootstrap process.

The patch also implements the fallback process in the pip_install role to
ensure that the fall back also takes place when the playbooks are run.

Change-Id: I33de85feacb633c10ed828f5d005992da6ca8a1e