Fix to correctly set the nova_management_address
This change sets the nova management address based on the
facts the play is able to set about the environment.

This will correct configure the nova to use a specific
management network which is not always the default
container address.

Closes-Bug: #1504226
Change-Id: Ib6d84072d379118ebfc560f4594c28e1fecf3b40
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit c506dda61d303735189abbee7c70d972b3096a40)
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>