Use pip install --proxy when $HTTPS_PROXY is set
During execution of our bootstrap scripts, we need to install some
packages before we have constructed any of the infrastructure. In
this case when installing behind a proxy we should respect that
proxy.

When $HTTPS_PROXY is not set but $HTTP_PROXY is, that will be used
instead. The use of trusted-host or other pip configurations to
enable this usage are out of scope and can be resolved separately.

Change-Id: I3df5ef6a7d33a3357e8c00f138b019c06c6bc3d6
Closes-Bug: 1504405
(cherry picked from commit a875596ba2a0c37fff131ff6ea1f899732b0f5ce)