Put swift {account,container,object} logs where they belong
The pattern for nova and neutron on hosts is to have a symlink from
/var/log/{service} to /openstack/log/_hostname_-{service}/ and then
to have all the service logs configured to log to /var/log/{service}
as that is a logical place for an operator to look for them.

Swift currently does not follow that pattern.

Currently the swift {account,container,object} logs are placed in
/openstack/log/{hostname}/, whereas the proxy logs are placed in
/var/log/swift/. On hosts the /var/log/swift symlink to
/openstack/log/{hostname}-swift/ is created, but not used.

This creates confusion for operators trying to find the logs in the
logical (and upstream) pattern in the directory /var/log/{service}.

This patch puts the swift logs where they belong.

Upgrade Notes:
 - This changes the location of the log storage on swift hosts from
   /openstack/log/{hostname} to /openstack/log/{hostname}-swift
 - Any log processing or monitoring tooling that consumes swift logs
   will need to be adjusted to consume them from the new location,
   or simply to consume them from /var/log/swift

DocImpact
UpgradeImpact
Closes-Bug: #1417536
Change-Id: I8d6ec98d310ce8d4e4a7a6cc5fb2d349d17757cf