Update heat keystone_authtoken config
This patch implements a slight re-ordering of some configuration in
order to match the pattern used in other roles (for easier
comparison). It also updates the keystone_authtoken configuration
to a more up to date set of configuration entries.

Upgrade Notes:
In order to standardise the location of data in the 'signing_dir'
(which is only used when PKI tokens are implemeted), the default
location of this data has been changed from '/var/cache/heat' to
'{{ heat_system_home_folder }}/cache/heat'. By default this
translates to '/var/lib/heat/cache/heat'.

DocImpact
UpgradeImpact
Implements: blueprint liberty-release
Change-Id: I7d3768a5629bb202883e82fe0c9c52fb0ca38b97
Co-Authored-By: Miguel Grinberg <miguelgrinberg50@gmail.com>