Store RabbitMQ cert/key in config dir
Storing rabbit's private key in /etc/ssl/private causes problems since that
directory (and the files within it) can only be accessed by root on Ubuntu
systems. Storing the key within the RabbitMQ configuration directory would
allow the key to be read by the 'rabbitmq' user.

The key can also be set to mode 0600 as well by moving its location and
changing it to be owned by the rabbitmq user.

Closes-bug: 1506992

Change-Id: Iede0748b57a86b33879d759505dd8f80476b574c