Ensure that the vnc console is disabled when spice is enabled
The spice and vnc consoles are mutually exclusive with Horizon set
to auto-detect the console type. It is also clear from the
configuration method that the designed implementation is to deploy
only one of them.

This patch ensures that only one is enabled at a time by explicitly
disabling the vnc console (which defaults to enabled).

Change-Id: Ib0368d3d6125e4abe7f9a9c40748457fc0d349f2