Add theme fix for browsable source code
This adds a fix to ensure that the published docs have the correct
link to the browsable source code as implemented in
https://review.openstack.org/236091

It also includes an exclusion of the doc directory from the pep8
check as the documentation conf.py requires the lengthy line.

It also updates the dev-requirements for the docs jobs to ensure
that the requirements are in line with upstream.

Change-Id: I52b5ede12667cf454d99a053e20eba4461b8ee75