Make the container cache resolvers configurable
This change simply makes it possible for a deployer to change
the in container resolvers.

In the bootstrap-aio script the resolvers are being set in the
user_variables so containers are using the same resolvers as the
host.

Change-Id: Id7ab3632217c3eb2217a308d7226736a67900da5
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit 8d51f7bafc38f69c493d8573b86ed8a0f629c463)