Updated the repo build templates
The change makes it so the upper constraints file is created
specifically using the repos indexed and cloned. Previously the
constraint file was created using the git repos found within the
local openstackgit directory and while that works for greenfield
deployments it does have the potential to create additional entries
that may or may not work with an environment.

The manifest file now uses normalized file names and will track the
exact git repos needed to power a build. This will allow for more
targetted syncs from an upstream mirror to happen using the
`repo-clone-mirror.yml` play.

The venv build process will now check for a prebuilt venv using the
versioned file name.

Change-Id: If4416f660133a6c65badba0c7c819a6409bfe11e
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>