Update juno for new dev work - 2 Nov 2015
Updates all repo SHAs to open up work on 10.1.17

This patch also implements an appropriate cap for the elasticsearch
python low level client to match the version of elasticsearch used,
and implements a cap for elasticsearch-curator for the same reason.

As the caps have been implemented, there is no longer a need to carry
the git sources too, so they've been removed.

The changes for elasticsearch were implemented as the latest versions
were being pulled in by the installation, causing problems with juno
build requirements and also causing the elasticsearch installation to
fail entirely.

Change-Id: Ia878d14c9f50869654aa2874e514930cec54e58f