Deployment fix when container_networks are not set
The change updates the logic within nova such to it will
now correctly set the management interface when using
a dynamic inventory without the container_networks options
within it.

This also updates the container create play so that it too
will properly skip a host that also may not have the container
networks entries within it's inventory configs.

Change-Id: I157a6ecd0ad99000369d786a649c7b40d791d917
Closes-Bug: #1508517
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>