Fixed haproxy backend config issue
The change updates the haproxy configs to use backends with a
default of an empty array. This is needed to ensure that the backends
are not being assumed to have some value. In the current system if an
inventory item did not exist the rendered backend would be a jinja
string which would cause the haproxy play to fail.

Closes-Bug: #1513455

Change-Id: I221dbd910c75c2e41c98a33e1c38c7f69718b544
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>