Fix issue with system installed xtrabackup
The change adds a task to ensure that the system provided
xtrabackup package is removed before installing the downloaded
version using the deb. This is required to ensure that there
are no version conflicts when installing the package.

The change also corrects the version of the galera server repo
in the defaults directory. While the repo correct version of
mariadb is being installed the name of the assumed repo is wrong.

Closes-Bug: #1513459

Change-Id: I63cb546cf5a4386320f01ed6f2120dd7c8341335
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>