Adding docs/comments for glance config
Deployers who use a dollar sign ($) in their swift backend passwords for glance
will be unable to authenticate to the swift backend. This is due to how
oslo.config formats strings.  This patch adds documentation and comments to
the user_variables.yml to ensure deployers know how to escape those strings.

This patch also adds a commented variable for the auth version on swift
backends so that deployers are reminded to use the correct version number.

Closes-bug: 1514291

Change-Id: I717e9afb5af1160de67dc5d0f87985fde5ad393e