Only use GET_PIP_URL when it is set
The logic in scripts/scripts-library.sh was recently changed with
respect to installing pip, however the conditional wasn't updated to
only use GET_PIP_URL when it is set.  This simple commits flips the
if statement to use a '-n' (checking if GET_PIP_URL is non-empty
instead) of a '-z'.

Change-Id: I6ace12f183e944a6cee2d7ab722424152f5eb419