Dropping logrotate.conf for all Openstack aggregate logs
This fix will add a logrotate configuration file for each host
including all log files found under /openstack/log.
All log files are rotated using copytruncate, so no service needs to
be restarted after the rotation.
This was implemended to mimic the kilo implementation, hence using the same
file name for the logrotate configuration, so it'll updated as part of the
kilo upgrade.

Change-Id: Ifb180bf5e1dd17c5ee3bdbfda54f7b469d1e4daf
Closes-Bug: #1516697