Add container-reconciler and object-expirer to os-swift
The container-reconciler and object-expirer were missing from the os-swift
role. The reconciler  makes sure incorrectly placed objets live in the
correct storage policy. The expirer is the service that deletes expired objects.

This change also adds the abilty to optionally specify a reclaim_age in the swift
section of the configuration, which is now set in all the locations required,
still with the default of 604800 seconds (7 days).

Change-Id: Ic56a714c3fb3c84b9bb5ed8e2ae3c86dad474161
Closes-Bug: #1516877
(cherry picked from commit b8fa505969c1ad7d38a83d6d87c9891326e20622)