Update requirements to suit the existence of lxc vg
If lxc volume group exists, the aio will try to use
this volume group to create the containers.

As one container will use a hard 5GB of storage, and
the fact that (by default) there is more than 16
containers, you'll exhaust all your storage space
and the aio build will fail.

This commit should make this failure more clear to
openstack-ansible users.

The amount of disk is counted with the default
install.

Change-Id: I84f68a24423f9e7046b1235ef51bf2291708fcec