Change AIO bootstrap disk preparation
This implements a change in the AIO disk preparation in the hope that
it will result in less Gate failures (due to the OS disk running low)
and will increase the flexibility slightly for dev/test purposes.

This introduces the following new environment variables:

 - BOOTSTRAP_AIO_DIR
   A data directory for the AIO's loopback disks. This was previously
   hard coded to '/opt' and is now set to '/openstack'.
 - DATA_DISK_DEVICE
   If there is a large enough secondary disk available, the bootstrap
   script will repartition it, format it and mount it at
   BOOTSTRAP_AIO_DIR.
   If you wish to specify which disk to use, then this var may be set.

This also changes the following environment variable:
 - MIN_LXC_VG_SIZE_GB -> MIN_DISK_SIZE_GB
   This represents the smallest disk size required (in GB) which may
   be repurposed by the bootstrap script. This size is set to 80G by
   default.

This patch also makes some minor optimisations in some of the grep
statements relating to the loopback disks and moves some of the gate
check script actions to accommodate the reparitioning of /openstack
which is where the gate check logs are held.

The new process implements the following major changes:

 - The BOOTSTRAP_AIO_DIR contains the loopback disks for
   cinder-volume, swift, swap and nova.
 - /var/lib/lxc will always be a normal ext4 partition - it used to
   be done this way in RAX Cloud and setup as a LVM VG in HP Cloud.
 - LXC containers will always use sparse files, unless the end-user
   sets up their own lxc vg.
 - The nova instances will now be built onto a sparse loopback disk
   mounted at /var/lib/nova/instances. This is specifically due to
   issues with that directory being on the root disk for gate tests
   and the root disks are too small to support instance creation.
 - The /openstack directory will be setup as a partition on the
   second disk. As this directory holds variable data this is deemend
   more appropriate for long running AIO testing and development.

Implements: blueprint gate-split
Change-Id: I96b0c787ed4b6bb01a629a7899bea581fb39be4f
(cherry picked from commit 3a31ac6af52d940bfdee3442b1e1dba260655008)